% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elf_summary_stats.R
\name{elf_summary_stats}
\alias{elf_summary_stats}
\title{Get Evaluative Linguistic Framework statistics for texts}
\usage{
elf_summary_stats(texts, parser = "udpipe")
}
\arguments{
\item{texts}{Vector of character strings}

\item{parser}{Tool used to tag parts of speech in the text. One of
"udpipe" or "spacy"}
}
\value{
Table of summary statistics for each text,
    with columns

    \itemize{
        \item \code{n_words} Number of words
        \item \code{n_content} Number of content words
        \item \code{n_clauses} Number of clauses (estimated using number of verbs)
        \item \code{prop_content} Proportion of content words (0.0 to 1.0)
        \item \code{prop_function} Proportion of function words (0.0 to 1.0)
        \item \code{simple_density} Approximate lexical density, \code{prop_content} * 10
        \item \code{lexical_density} Lexical density, \code{prop_content} / \code{n_clauses}
    }
}
\description{
Get Evaluative Linguistic Framework statistics for texts
}
\examples{
elf_summary_stats(c("The quick brown fox jumped swiftly over the lazy dog.",
                    "She told him that she loved him."))
}
