% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_tags.R
\name{get_pos_tags}
\alias{get_pos_tags}
\title{Get part of speech tags}
\usage{
get_pos_tags(texts)
}
\arguments{
\item{texts}{vector of character strings}
}
\value{
Dataframe with tags for each word in the texts
}
\description{
Get part of speech tags
}
\examples{
get_pos_tags(c("The quick brown fox jumped swiftly over the lazy dog.",
               "She told him that she loved him."))
}
