% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_type.R
\name{get_pos_type}
\alias{get_pos_type}
\title{Identify content vs. function words}
\usage{
get_pos_type(tag)
}
\arguments{
\item{tag}{Vector of part of speech tags}
}
\value{
Character vector containing "Content", "Function", or NA
}
\description{
Identify content vs. function words
}
