% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spacy_tags.R
\name{get_spacy_tags}
\alias{get_spacy_tags}
\title{Get part of speech tags using spacy}
\usage{
get_spacy_tags(texts)
}
\arguments{
\item{texts}{vector of character strings}
}
\value{
tibble with tags for each word/token in the texts
}
\description{
Get part of speech tags using spacy
}
