% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_udpipe_tags.R
\name{get_udpipe_tags}
\alias{get_udpipe_tags}
\title{Get part of speech tags using the udpipe tagger}
\usage{
get_udpipe_tags(texts)
}
\arguments{
\item{texts}{vector of character strings}
}
\value{
tibble with tags for each word in the texts
}
\description{
Get part of speech tags using the udpipe tagger
}
