% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tagger.R
\name{load_tagger}
\alias{load_tagger}
\title{Load the udpipe English model}
\usage{
load_tagger()
}
\value{
A udpipe model, ready to be used in \code{\link[udpipe]{udpipe_annotate}}.
}
\description{
Load the udpipe English model
}
