% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spacy_tags.R
\name{get_spacy_tags}
\alias{get_spacy_tags}
\title{Get part of speech tags using spacy}
\usage{
get_spacy_tags(texts)
}
\arguments{
\item{texts}{vector of character strings}
}
\value{
Dataframe with tags for each word/token in the texts
}
\description{
Get part of speech tags using spacy
}
\examples{
get_spacy_tags(c("The quick brown fox jumped swiftly over the lazy dog.",
                 "She told him that she loved him."))
}
